Move dao functions into their own class
Move the functions that are responsible for connecting to the database
into their own class. In addition, refactor the resource class to be a
little bit neater about serialization and deserialization of request /
response JSON. Finally, bump the version number for JUnit because their
was a version bump on their side, and the plugin was pulling in bad
metadata now.
